initial refactoring and abstract annotated argument binder
